requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
plotly==5.18.0 
pandas==2.1.4